♻️ Refactor : 파일 분리, 내 히스토리 패널 버튼 수정
