+++ src/utils/authSession/constants.ts
@@ -1,0 +1,3 @@
+export const ACCESS_TOKEN_COOKIE_KEY = 'access-token';
+export const AUTH_SESSION_STORAGE_KEY = 'coworkers-auth-session';
+export const AUTH_SESSION_CHANGE_EVENT = 'coworkers-auth-session-change';
